moved interchage adaptor tests to birth and death specific modules
