+++ nhais-adaptor/adaptor/outgoing/death/tests/fixtures.py
@@ -6,9 +6,13 @@ from adaptor.outgoing.common.tests.fixtures import create_simple_practitioner
 
 
 def create_operation_definition_for_death_registration():
+    op_param_interchange_sequence = creators.create_parameter_with_binding(
+        name=ParameterName.INTERCHANGE_SEQ_NO,
+        value="000001")
     op_param_message_sequence = creators.create_parameter_with_binding(name=ParameterName.MESSAGE_SEQ_NO,
                                                                        value="000001")
     op_param_nhais_cypher = creators.create_parameter_with_binding(name=ParameterName.NHAIS_CYPHER, value="XX1")
+    op_param_sender_cypher = creators.create_parameter_with_binding(name=ParameterName.SENDER_CYPHER, value="TES5")
     op_param_transaction_number = creators.create_parameter_with_binding(name=ParameterName.TRANSACTION_NO,
                                                                          value="17")
     practitioner = create_simple_practitioner()
@@ -27,9 +31,11 @@ def create_operation_definition_for_death_registration():
     op_def = creators.create_operation_definition(name=OperationName.REGISTER_DEATH, code="gpc.registerpatient",
                                                   date_time="2019-04-23 09:00:04.159338",
                                                   contained=[practitioner, patient],
-                                                  parameters=[op_param_message_sequence,
+                                                  parameters=[op_param_interchange_sequence,
+                                                              op_param_message_sequence,
                                                               op_param_transaction_number,
                                                               op_param_nhais_cypher,
+                                                              op_param_sender_cypher,
                                                               op_param_practitioner, op_param_patient])
 
     return op_def
