+++ nhais-adaptor/mailbox/NHAIS/HA01/inbox/GP05-000002.txt
@@ -1,0 +1,15 @@
+UNB+UNOA:2+GP05+HA01+190423:0900+000002++FHSREG'
+UNH+000003+FHSREG:0:1:FH:FHS001'
+BGM+++507'
+NAD+FHS+HA:954'
+DTM+137:201904230900:203'
+RFF+950:G5'
+S01+1'
+RFF+TN:19'
+NAD+GP+4826940,281:900'
+GIS+1:ZZZ'
+DTM+961:20190420:102'
+S02+2'
+PNA+PAT+NHSNO222222:OPI'
+UNT+13+000003'
+UNZ+1+000002'
